+++ lib/components/document_details.dart
@@ -1,8 +1,9 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:szikapp/ui/themes.dart';
 import 'package:url_launcher/url_launcher.dart';
+
 import '../../models/goodtoknow.dart';
+import '../ui/themes.dart';
 
 class DocumentDetails extends StatelessWidget {
   final GoodToKnow? document;
