+++ pubspec.yaml
@@ -2,7 +2,7 @@ name: szikapp
 description: Awesome application for doing business in da SZIK.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-version: 0.0.1+57
+version: 0.0.1+58
 
 environment:
   sdk: ">=2.13.0 <3.0.0"
